requirements.txt
wandb
torch
torchvision
pylint
pillow
timm
